chore(lib-storage): move test files to src folder
